pyproject.toml
[tool.poetry]
name = "novizi"
version = "0.1.0"
description = "event management and ticketing service. Allows users to browse, create, and promote local events."
authors = ["Strategy Tap <strategy-tap@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "novizi.com"
documentation = ""
repository = "https://github.com/Strategy-Tap/novizi/"
keywords = ["novizi"]

[tool.poetry.dependencies]
python = "3.8.3"
argon2-cffi = "^20.1.0"
confusable_homoglyphs = "^3.2.0"
dj-database-url = "^0.5"
dj-rest-auth = {extras = ["with_social"], version = "^1.0"}
django = "^3.0"
djangorestframework = "^3.11"
djangorestframework-simplejwt = "^4.4.0"
django-axes = "^5.3"
django-cors-headers = "^3.3"
django-filter = "^2.3"
django-import-export = "^2.2"
django-storages = {extras = ["dropbox"], version = "^1.9"}
django-simple-history = "^2.10.0"
drf-yasg = "^1.17"
gunicorn = "^20.0"
Pillow = "^7.1"
psycopg2-binary = "^2.8"
python-decouple = "^3.3"
pwned-passwords-django = "^1.4"
whitenoise = {extras = ["brotli"], version = "^5.1"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = {extras = ["toml"], version = "^5.1"}
django-coverage-plugin = "^1.8"
django-debug-toolbar = "^2.2"
darglint = "^1.4.0"
flake8 = "^3.8.2"
flake8-annotations = "^2.1.0"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
isort = {extras = ["pyproject"], version = "^4.3.21"}
mypy = "^0.780"
mkdocs = "^1.1"
mkdocs-material = "^5.2"
mkdocstrings = "^0.11.3"
mkdocs-minify-plugin = "^0.3.0"
mkdocs-git-revision-date-localized-plugin = "^0.5.2"
pytest = "^5.4.2"
pytest-cov = "^2.9.0"
pytest-django = "^3.9"
radon = "^4.1"
safety = "^1.9.0"
xdoctest = "^0.12.0"

[tool.coverage.run]
branch = true
source = ["novizi"]
plugins = "django_coverage_plugin"

[tool.coverage.report]
show_missing = true
fail_under = 84

[tool.isort]
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 80

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
